requirements.txt
pillow>8.4.0
pytrends>4.7.3
pandas==1.2.5
tweepy>4.4.0
